Exercices/requirements.txt
Package            Version
------------------ -------
bcrypt             4.0.1
blinker            1.6.2
cffi               1.15.1
click              8.1.3
cryptography       40.0.2
fabric             3.0.0
Flask              2.3.1
Flask-SQLAlchemy   3.0.3
greenlet           2.0.2
importlib-metadata 6.6.0
invoke             2.1.0
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.2
paramiko           3.1.0
pip                22.1.1
pycparser          2.21
PyNaCl             1.5.0
python-dotenv      1.0.0
setuptools         62.3.2
SQLAlchemy         2.0.11
typing_extensions  4.5.0
Werkzeug           2.3.1
zipp               3.15.0
